Added new NPC
New NPC: "Spectral Malevolence"

This NPC is DANGEROUS because its invisible. If the room is dark, you have no chance to see him. If there is light in the room, you will see something like a moving steam, very transparent. To be able to properly see it, you need the flashlight or the torch. With this update, I managed to make the flashlight not just some tool to see in the dark, but actually something you need to survive. Now players will spend time to search for batteries.